feat(storage): add function to print responses
Despite the log message, the standard logger also provides date and
time. However, if for example, a subsystem is created, the desirable output is
just subsystem name, since it will simplify scripting. log package is
also used in tests and date/time can be useful there. As a solution
printResponse function is added to provide only required output.

Signed-off-by: Artsiom Koltun <artsiom.koltun@intel.com>